### 2017-12-15 #509 @whedon
[PRE REVIEW]: tidyhydat: Extract and Tidy Canadian Hydrometric Data
**Submitting author:** @boshek (<a href="http://orcid.org/0000-0002-9270-7884">Samuel Albers</a>)
**Repository:** <a href="https://github.com/ropensci/tidyhydat" target ="_blank">https://github.com/ropensci/tidyhydat</a>
**Version:** 0.3.0
