### 2017-12-15 #509 @arfon
@openjournals/ropensci - has this package already on-boarded? If so, could you share the thread please?
