### 2017-12-15 #509 @arfon
@boshek - please merge this PR so we can proceed https://github.com/ropensci/tidyhydat/pull/69
