### 2017-12-15 #509 @boshek
@arfon Merged. Thanks!
