### 2017-12-15 #511 @arfon
Accepting because of rOpenSci acceptance in https://github.com/ropensci/tidyhydat/pull/69
