### 2017-12-15 0f12510
Minor edits for JOSS submission
