+++ inst/JOSS/paper.bib
@@ -35,7 +35,7 @@
   title={R for data science},
   author={Wickham, Hadley and Grolemund, Garrett},
   year={2016},
-  publisher={Sebastopol, CA: O’Reilly. http:// r4ds. had. co. nz}
+  publisher={Sebastopol, CA: O'Reilly. http://r4ds.had.co.nz}
 }
 
 @Manual{RCore,
@@ -55,12 +55,10 @@
   url = {https://CRAN.R-project.org/package=tidyverse},
 }
 
-  @Manual{alberstidyhydat,
-    title = {tidyhydat: Extract and Tidy Canadian Hydrometric Data},
-    author = {Sam Albers},
-    year = {2017},
-    note = {R package version 0.3.0},
-    url = {https://github.com/ropensci/tidyhydat},
-  }
-
-@Comment{jabref-meta: databaseType:bibtex;}
+@Manual{alberstidyhydat,
+  title = {tidyhydat: Extract and Tidy Canadian Hydrometric Data},
+  author = {Sam Albers},
+  year = {2017},
+  note = {R package version 0.3.0},
+  url = {https://github.com/ropensci/tidyhydat},
+}
