### 2017-12-15 #511 @arfon
@boshek - one more set of small changes https://github.com/ropensci/tidyhydat/pull/70
