### 2017-12-15 #511 @boshek
@arfon Merged
