### 2017-12-15 #511 @arfon
@boshek your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00511 ⚡️ 🚀 💥
