### 2018-01-22 #511 @boshek
@arfon Embarrassing ask: Is there any opportunity to correct a few spelling errors in the article?
