### 2018-01-22 #511 @arfonsmith
Sure thing. Please make the edits and I can update the paper.
