### 2018-01-24 #511 @boshek
Thanks @arfon . Edits made here: https://github.com/ropensci/tidyhydat/blob/master/inst/JOSS/paper.md
